+++ plotly/src/layout/mod.rs
@@ -17,6 +17,7 @@ mod axis;
 mod geo;
 mod grid;
 mod legend;
+mod map;
 mod mapbox;
 mod modes;
 mod polar;
@@ -40,6 +41,7 @@ pub use self::axis::{
 pub use self::geo::LayoutGeo;
 pub use self::grid::{GridDomain, GridPattern, GridXSide, GridYSide, LayoutGrid, RowOrder};
 pub use self::legend::{GroupClick, ItemClick, ItemSizing, Legend, TraceOrder};
+pub use self::map::{LayoutMap, MapBounds, MapStyle};
 pub use self::mapbox::{Center, Mapbox, MapboxStyle};
 pub use self::modes::{
     AspectMode, BarMode, BarNorm, BoxMode, ClickMode, UniformTextMode, ViolinMode, WaterfallMode,
@@ -343,6 +345,7 @@ pub struct LayoutFields {
     // ternary: Option<LayoutTernary>,
     scene: Option<LayoutScene>,
     geo: Option<LayoutGeo>,
+    map: Option<LayoutMap>,
     polar: Option<LayoutPolar>,
     annotations: Option<Vec<Annotation>>,
     shapes: Option<Vec<Shape>>,